fix: Override-only API violation
